gcp/src/client/requirements.txt
protobuf==4.25.5
fastapi==0.104.1
uvicorn[standard]==0.24.0
loguru==0.7.2
joblib==1.3.2
pandas==2.1.3
numpy==1.24.4            
mlflow==2.8.1
pydantic==1.10.8
ortools==9.7.2996        
xgboost==2.1.4
streamlit
requests
python-dotenv
scikit-learn 
optbinning
boto3
requests
urllib3
minio
evidently
lightgbm
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-asgi
prometheus_client
opentelemetry-instrumentation-system-metrics